refactor: remove duplication in type declaration
