+++ packages/agent/src/types.ts
@@ -1,10 +1,5 @@
-import {
-  CollectionActionEvent,
-  GenericTree,
-  SmartActionApprovalRequestBody,
-  SmartActionRequestBody,
-} from '@forestadmin/forestadmin-client';
 import { CompositeId, Logger, LoggerLevel } from '@forestadmin/datasource-toolkit';
+import { ForestAdminClient } from '@forestadmin/forestadmin-client';
 import { IncomingMessage, ServerResponse } from 'http';
 
 /** Options to configure behavior of an agent's forestadmin driver */
@@ -50,38 +45,3 @@ export type SelectionIds = {
   areExcluded: boolean;
   ids: CompositeId[];
 };
-
-export type User = Record<string, any> & {
-  id: number;
-  tags: Record<string, string>;
-};
-
-export interface ForestAdminClient {
-  canOnCollection(params: {
-    userId: number;
-    event: CollectionActionEvent;
-    collectionName: string;
-  }): Promise<boolean>;
-  canExecuteCustomAction(params: {
-    userId: number;
-    customActionName: string;
-    collectionName: string;
-    body: SmartActionRequestBody | SmartActionApprovalRequestBody;
-  }): Promise<false | SmartActionRequestBody>;
-  canExecuteCustomActionHook(params: {
-    userId: number;
-    collectionName: string;
-    customActionName: string;
-  }): Promise<boolean>;
-  getScope(params: {
-    renderingId: number;
-    user: User;
-    collectionName: string;
-  }): Promise<GenericTree>;
-  canRetrieveChart(params: {
-    renderingId: number;
-    userId: number;
-    chartRequest: SmartActionApprovalRequestBody | SmartActionRequestBody;
-  }): Promise<boolean>;
-  markScopesAsUpdated(renderingId: number);
-}
