+++ packages/forestadmin-client/package.json
@@ -32,6 +32,7 @@
     "@types/jsonwebtoken": "^8.5.8",
     "@types/lru-cache": "^7.6.1",
     "@types/superagent": "^4.1.15",
-    "fishery": "^2.2.2"
+    "fishery": "^2.2.2",
+    "@forestadmin/datasource-toolkit": "1.0.0-alpha.3"
   }
 }
